code review changes done

+++ src/main/java/com/example/demo/model/Telephone.java
@@ -19,5 +19,5 @@ public class Telephone {
   private TelephoneCode telephoneCode;
 
   @Column(name = "telephone")
-  private int telephoneNumber;
+  private Long telephoneNumber;
 }
