+++ src/main/java/com/example/demo/model/dto/TelephoneDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 @Builder
 public class TelephoneDTO {
-    private Long id;
-    private TelephoneCode telephoneCode;
-    private int telephoneNumber;
+  private Long id;
+  private TelephoneCode telephoneCode;
+  private Long telephoneNumber;
 }
